Move the enhance_pipe_sections_appearance method to the Model class
