+++ pulse/model/model.py
@@ -9,6 +9,10 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from pulse.project.project import Project
 
+from pulse.model.acoustic_element import AcousticElement
+from pulse.model.structural_element import StructuralElement
+from pulse.model.cross_sections.expansion_joint_cross_section import ExpansionJointCrossSection
+
 import numpy as np
 
 class Model:
@@ -173,3 +177,79 @@ class Model:
 
         # if self.list_frequencies != frequencies:
         #     return True
+
+    def enhance_pipe_sections_appearance(self):
+        """ 
+        This method adds lids to cross-section variations and terminations.
+        """
+        for elements in self.preprocessor.structural_elements_connected_to_node.values():
+
+            element = None
+            if len(elements) == 2:
+
+                first_element, last_element = elements
+                first_element: StructuralElement
+                last_element: StructuralElement
+
+                if 'beam_1' in [first_element.element_type, last_element.element_type]:
+                    continue
+
+                first_cross = first_element.cross_section
+                last_cross = last_element.cross_section
+                
+                if (first_cross, last_cross).count(None):
+                    continue
+
+                first_outer_diameter = first_cross.outer_diameter
+                first_inner_diameter = first_cross.inner_diameter
+                last_outer_diameter = last_cross.outer_diameter
+                last_inner_diameter = last_cross.inner_diameter
+
+                if first_outer_diameter < last_inner_diameter:
+                    inner_diameter = first_inner_diameter 
+                    element = last_element
+
+                if last_outer_diameter < first_inner_diameter:
+                    inner_diameter = last_inner_diameter 
+                    element = first_element
+
+            elif len(elements) == 1: 
+
+                element = elements[0]
+                element: StructuralElement
+ 
+                if element.element_type == 'beam_1':
+                    continue  
+
+                first_node = element.first_node
+                last_node = element.last_node  
+
+                if element.cross_section is None:
+                    continue
+
+                inner_diameter = element.cross_section.inner_diameter 
+
+                if len(self.preprocessor.neighbors[first_node]) == 1:
+                    first_node_id = first_node.external_index
+                    if self.properties.is_there_an_acoustic_attribute_in_the_node(first_node_id) == 0:
+                        inner_diameter = 0
+
+                elif len(self.preprocessor.neighbors[last_node]) == 1:
+                    last_node_id = last_node.external_index
+                    if self.properties.is_there_an_acoustic_attribute_in_the_node(last_node_id) == 0:
+                        inner_diameter = 0
+
+            if isinstance(element, AcousticElement | StructuralElement):
+
+                if element.element_type == 'expansion_joint':
+                    section_info = element.cross_section.section_info
+                    if isinstance(section_info, ExpansionJointCrossSection):
+                        element.section_parameters_render = section_info._as_list()
+
+                else:
+
+                    section_info = element.cross_section.section_info
+                    outer_diameter, _, offset_y, offset_z, t_ins, *_ = section_info.section_parameters
+
+                    thickness = (outer_diameter - inner_diameter) / 2
+                    element.section_parameters_render = [outer_diameter, thickness, offset_y, offset_z, t_ins]
