+++ pulse/project/project.py
@@ -4,8 +4,6 @@ import logging
 from collections import defaultdict
 from pathlib import Path
 
-import numpy as np
-
 from pulse import TEMP_PROJECT_DIR, app
 from pulse.editor import Pipeline
 from pulse.interface.file.project_file import ProjectFile
@@ -18,10 +16,6 @@ from pulse.model.model import Model
 from pulse.processing.acoustic_solver import AcousticSolver
 from pulse.processing.structural_solver import StructuralSolver
 from pulse.project.load_project import LoadProject
-from pulse.model.acoustic_element import AcousticElement
-from pulse.model.structural_element import StructuralElement
-from pulse.model.cross_sections.expansion_joint_cross_section import ExpansionJointCrossSection
-from pulse.model.cross_sections.valve_cross_section import ValveCrossSection
 
 
 error_title = "Error"
@@ -190,98 +184,6 @@ class Project:
         app().main_window.update_status_bar_info()
         # dt = time()-t0
         # print(f"Time to process_geometry_and_mesh: {dt} [s]")
-
-    def enhance_pipe_sections_appearance(self):
-        """ 
-        This method adds lids to cross-section variations and terminations.
-        """
-        for elements in self.model.preprocessor.structural_elements_connected_to_node.values():
-
-            element = None
-            if len(elements) == 2:
-                first_element, last_element = elements
-                first_element: StructuralElement
-                last_element: StructuralElement
-                
-                if 'beam_1' not in [first_element.element_type, last_element.element_type]:
-                    first_cross = first_element.cross_section
-                    last_cross = last_element.cross_section
-                    
-                    if not (first_cross and last_cross):
-                        continue
-
-                    first_outer_diameter = first_cross.outer_diameter
-                    first_inner_diameter = first_cross.inner_diameter
-                    last_outer_diameter = last_cross.outer_diameter
-                    last_inner_diameter = last_cross.inner_diameter
-
-                    if first_outer_diameter < last_inner_diameter:
-                        inner_diameter = first_inner_diameter 
-                        element = last_element
-
-                    if last_outer_diameter < first_inner_diameter:
-                        inner_diameter = last_inner_diameter 
-                        element = first_element
-
-            elif len(elements) == 1: 
-
-                element = elements[0]
-                element: StructuralElement
- 
-                if element.element_type == 'beam_1':
-                    continue  
-
-                first_node = element.first_node
-                last_node = element.last_node  
-
-                if element.cross_section is None:
-                    continue
-
-                inner_diameter = element.cross_section.inner_diameter 
-
-                if len(self.model.preprocessor.neighbors[first_node]) == 1:
-                    first_node_id = first_node.external_index
-                    if self.is_there_an_acoustic_attribute_in_the_node(first_node_id) == 0:
-                        inner_diameter = 0
-
-                elif len(self.model.preprocessor.neighbors[last_node]) == 1:
-                    last_node_id = last_node.external_index
-                    if self.is_there_an_acoustic_attribute_in_the_node(last_node_id) == 0:
-                        inner_diameter = 0
-
-            if isinstance(element, AcousticElement | StructuralElement):
-
-                if element.element_type == 'expansion_joint':
-                    section_info = element.cross_section.section_info
-                    if isinstance(section_info, ExpansionJointCrossSection):
-                        element.section_parameters_render = section_info._as_list()
-
-                else:
-
-                    section_info = element.cross_section.section_info
-                    outer_diameter, _, offset_y, offset_z, t_ins, *_ = section_info.section_parameters
-
-                    thickness = (outer_diameter - inner_diameter) / 2
-                    element.section_parameters_render = [outer_diameter, thickness, offset_y, offset_z, t_ins]
-
-    def is_there_an_acoustic_attribute_in_the_node(self, node_id: int):
-
-        acoustic_properties = [
-            "acoustic_pressure", 
-            "volume_velocity", 
-            "specific_impedance", 
-            "radiation_impedance", 
-            "reciprocating_compressor_excitation",
-            "reciprocating_pump_excitation",
-            "psd_acoustic_link",
-            "acoustic_transfer_element"
-        ]
-
-        for (property, *args) in self.model.properties.nodal_properties.keys():
-            if property in acoustic_properties and node_id in args:
-                    return True
-
-        return False
 
     def is_analysis_setup_complete(self):
